color applier optim

+++ common/src/main/java/net/lerariemann/infinity/block/entity/NeitherPortalBlockEntity.java
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.Nullable;
 public class NeitherPortalBlockEntity extends BlockEntity {
     private final PropertyDelegate propertyDelegate;
     private Identifier dimension;
-    private long portalColor;
+    private int portalColor;
     private boolean isOpen;
 
     public NeitherPortalBlockEntity(BlockPos pos, BlockState state) {
@@ -22,7 +22,7 @@ public class NeitherPortalBlockEntity extends BlockEntity {
         this.propertyDelegate = new PropertyDelegate() {
             public int get(int index) {
                 if (index == 0) {
-                    return (int)(NeitherPortalBlockEntity.this.portalColor);
+                    return NeitherPortalBlockEntity.this.portalColor;
                 }
                 return 0;
             }
@@ -39,11 +39,11 @@ public class NeitherPortalBlockEntity extends BlockEntity {
         };
     }
 
-    public NeitherPortalBlockEntity(BlockPos pos, BlockState state, long i) {
+    public NeitherPortalBlockEntity(BlockPos pos, BlockState state, int i) {
         this(pos, state, i, InfinityMod.getDimId(i));
     }
 
-    public NeitherPortalBlockEntity(BlockPos pos, BlockState state, long i, Identifier id) {
+    public NeitherPortalBlockEntity(BlockPos pos, BlockState state, int i, Identifier id) {
         this(pos, state);
         this.dimension = id;
         this.portalColor = i;
@@ -53,7 +53,7 @@ public class NeitherPortalBlockEntity extends BlockEntity {
     public Identifier getDimension() {
         return this.dimension;
     }
-    public long getPortalColor() {
+    public int getPortalColor() {
         return this.portalColor;
     }
     public boolean getOpen() {
@@ -61,10 +61,10 @@ public class NeitherPortalBlockEntity extends BlockEntity {
     }
 
     public void setDimension(long c) {
-        setDimension(c, InfinityMod.getDimId(c));
+        setDimension((int)c, InfinityMod.getDimId(c));
     }
 
-    public void setDimension(long c, Identifier i) {
+    public void setDimension(int c, Identifier i) {
         this.portalColor = c;
         this.dimension = i;
     }
@@ -81,7 +81,7 @@ public class NeitherPortalBlockEntity extends BlockEntity {
 
     public void readNbt(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup) {
         super.readNbt(tag, registryLookup);
-        this.portalColor = tag.getLong("Dimension");
+        this.portalColor = tag.getInt("Dimension");
         if (tag.contains("DimensionName")) {
             this.dimension = Identifier.of(tag.getString("DimensionName"));
         }
